SmartTrader/requirements.txt
# Core Data
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Market Data (Free)
yfinance>=0.2.40
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP & Sentiment (Lightweight)
vaderSentiment>=3.3.2
textblob>=0.17.1
# Note: transformers and sentence-transformers in requirements-ml.txt

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
colorama>=0.4.6

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Logging
loguru>=0.7.0

# Dashboard
streamlit>=1.28.0

# Reddit API
praw>=7.7.0  # Fixed: was >=3.7.0 which could pull incompatible version
